Commit 9: The main screen has been refactored. Screen templates for general balance and world time have been created.
