+++ app/src/main/java/com/itskidan/currencyexchangeapp/ui/components/Advertising.kt
@@ -1,4 +1,4 @@
-package com.itskidan.currencyexchangeapp.ui.googleadd
+package com.itskidan.currencyexchangeapp.ui.components
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
