+++ app/src/main/java/com/itskidan/currencyexchangeapp/ui/screens/totalbalance/TotalBalanceViewModel.kt
@@ -1,0 +1,6 @@
+package com.itskidan.currencyexchangeapp.ui.screens.totalbalance
+
+import androidx.lifecycle.ViewModel
+
+class TotalBalanceViewModel : ViewModel() {
+}
