+++ app/src/main/java/com/itskidan/currencyexchangeapp/ui/screens/worldtime/WorldTimeViewModel.kt
@@ -1,0 +1,6 @@
+package com.itskidan.currencyexchangeapp.ui.screens.worldtime
+
+import androidx.lifecycle.ViewModel
+
+class WorldTimeViewModel:ViewModel() {
+}
